ros_control: restructer servo interface
