+++ bitbots_ros_control/include/bitbots_ros_control/utils.h
@@ -5,9 +5,18 @@
 #include "humanoid_league_msgs/Speak.h"
 
 namespace bitbots_ros_control {
-    void speakError(const ros::Publisher &speak_pub, std::string text);
-    uint16_t dxlMakeword(uint64_t a, uint64_t b);
-    uint32_t dxlMakedword(uint64_t a, uint64_t b);
+
+enum ControlMode {
+  POSITION_CONTROL,
+  VELOCITY_CONTROL,
+  EFFORT_CONTROL,
+  CURRENT_BASED_POSITION_CONTROL
+};
+
+bool stringToControlMode(std::string _control_modestr, ControlMode &control_mode);
+void speakError(const ros::Publisher &speak_pub, std::string text);
+uint16_t dxlMakeword(uint64_t a, uint64_t b);
+uint32_t dxlMakedword(uint64_t a, uint64_t b);
 }
 
 #endif  //BITBOTS_ROS_CONTROL_INCLUDE_BITBOTS_ROS_CONTROL_UTILS_H_
