+++ bitbots_ros_control/src/node.cpp
@@ -14,7 +14,7 @@ int main(int argc, char *argv[]) {
     return 1;
   }
 
-  ROS_WARN("%d", hw.getNames().size());
+  //ROS_WARN("%l", hw.getNames().size());
   for (std::string &hw_name : hw.getNames()) {
     ROS_WARN("%s", hw_name.c_str());
   }
